+++ Ruddarr/Views/ContentViewMac.swift
@@ -45,7 +45,6 @@ struct ContentView: View {
         .displayToasts()
         .whatsNewSheet()
         .reportBugSheet()
-        .instanceWebSheet()
         .onBecomeActive(perform: handleScenePhaseChange)
     }